requirements.txt
streamlit
pandas
numpy
langchain
faiss-cpu
sentence-transformers
python-dotenv
PyPDF2
python-docx
torch
google-generativeai
diskcache
types-requests